perf: remove lock in NioStatus pending compute

+++ client/src/main/java/com/github/dtprj/dongting/net/NioNet.java
@@ -55,7 +55,7 @@ public abstract class NioNet extends AbstractLifeCircle {
     private final Condition pendingReqCond = lock.newCondition();
     private final Condition pendingBytesCond = lock.newCondition();
     int pendingRequests;
-    int pendingBytes;
+    long pendingBytes;
 
     private final boolean server;
